Goblin: Eyes Started!
The head is now more rounded except for the back, which will need work. The nose and eyebrows have been fixed, and wrinkles have been added to the forehead. Eye sockets have been marked as well. The next step will be to round out the back of the head and complete it, and then on to fixing the mouth to look more like a mouth. Once that's done, the next step will be to open the eye socket and the mouth, and to start on the eyes and the teeth! (Note: The iris of the eye (the colored part) will be very large, and the majority of the eye will be black (much like the eye of a dog.))