Goblin: Head and Mouth Adjustments
Tongue is now added, and ears are now started. Skull has been adjusted and throat hole has been fixed. The next step is to complete the ears, and then start working on the teeth. Maybe play around with extruding the tongue to see how it looks when the Goblin sticks out his/her tongue!